pyproject.toml
[project]
name = "research-page"
version = "0.1.0"
description = "Research landing page microservice"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "bleach==6.2.0",
  "fastapi==0.115.12",
  "jinja2==3.1.6",
  "markdown==3.8",
  "pyyaml==6.0.2",
  "redis==5.2.1",
  "uvicorn[standard]==0.34.0",
]

[dependency-groups]
dev = [
  "httpx==0.28.1",
  "pytest==8.3.5",
  "ruff==0.11.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.per-file-ignores]
"tests/test_app.py" = ["E402"]
"tests/test_redis_backend.py" = ["E402"]
"tests/test_redis_cluster_backend.py" = ["E402"]

[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"
